requirements.txt
mss
pillow
pynput
PyTurboJPEG
opencv-python
dxcam #windows版本需要
pylibfec #windows fast 版本需要
av #windows fast 版本需要
psutil #windows fast 版本需要